Add CSS for scraps and printing
Use flexbox to format and distribute the scraps. Flexbox and min-height
make sure that all scraps in a row have the same height (as opposed to
float where they might cluster).

Printing the correct sizes does not work properly for Chrome and Safari due to this bug:
https://bugs.chromium.org/p/chromium/issues/detail?id=273306

TODO: Add help text that describes the different formats and the border
feature.