Add formatAsPasssword filter
columns can be formatted with {{ $1|formatAsPassword|safe }}

However this leads to invalid HTML if the password conations HTML-like
characters.

Instead, we should add the character analysis as a function to the
template data. It should return an array of {char:"x", type:"x"} object)
and iterate over the array in the template, using the automatic
escaping.